Startup on F#
Trying different options and learning the language have so far only
implemented the reading of the file and loading of the matrix as I was
playing around with the different functionality. I still need to add an
accelerator but I want to write the processor code first to gain an
understanding of F# and then convert the code to Accelerator. Progress
of this code seems a lot more smooth than Harlan.