Found a work around and added all the different tests but Harlan and CUDA
Found a work around and added all the different tests but Harlan and
CUDA. I Just need to recompile CUDA then add it to the project as the
required code has been added. Then the Automatic Testing can start. And
if you Use it remember to add the Matrix's from the RAR to the solution.
Found additional Errors in Kmeans-AMP which makes it next to impossible
to run in Automation.