AMP now partially works.
AMP now partially works. The code now works... however, I am unsure if
it runs it on multiple threads on the CPU or it runs it on the GPU or
both. However, It could not run only on the GPU as it gave errors. and
the previous way created a lot of errors and did not work as intended.